LeverObject.cs: lever now operates like a switch
