+++ Assets/Prefabs/Lever.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: d1245311ba05c90469fe3ce9965eb4c7
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
